Implement `getRecordById` in RecordDao
- Add a new function `getRecordById` to `RecordDao` to retrieve a record by its ID.
- The function uses a SQL query to select a record with a specified ID from the `records` table.
- Change type of the `recordId` parameter from `Int` to `Long`.

+++ app/src/main/java/com/ihff/fuelelectro/data/database/RecordDao.kt
@@ -24,5 +24,5 @@ interface RecordDao {
     suspend fun updateRecord(record: Record)
 
     @Query("SELECT * FROM records WHERE id = :recordId")
-    suspend fun getRecordById(recordId: Int): Record?
+    suspend fun getRecordById(recordId: Long): Record?
 }
